refactor: converte páginas para server components
Dashboard e transações passam a buscar dados no servidor via async/await
com Promise.all, eliminando useEffect/useState nas pages. Lógica client
da página de transações extraída para TransactionPageClient. DashboardMenu
usa usePathname para detectar rota ativa em vez de estado local. NewTransactionAction
passa a aceitar categories via prop para evitar fetch duplicado. DataTable
recebe tipagem correta para ColumnMeta do tanstack-table.
