+++ src/app/(main)/dashboard/components/dashboardMenu.tsx
@@ -1,33 +1,29 @@
 "use client";
 
 import Link from "next/link";
-import { useState } from "react";
+import { usePathname } from "next/navigation";
 import type { LucideIcon } from "lucide-react";
 import { CreditCard, Home } from "lucide-react";
 
-export default function DashboardMenu() {
-  const [selectedMenu, setSelectedMenu] = useState("home");
+const menuItems = [
+  { name: "Home", route: "/dashboard", icon: Home },
+  { name: "Relatórios Financeiros", route: "/em-construcao", icon: CreditCard },
+];
 
-  function checkSelected(key: string) {
-    return selectedMenu === key;
-  }
+export default function DashboardMenu() {
+  const pathname = usePathname();
 
   return (
     <section className="flex w-full flex-col gap-2 bg-white p-4">
-      <DashboardMenuItem
-        name="Home"
-        route="/dashboard"
-        icon={Home}
-        selected={checkSelected("home")}
-        onSelect={() => setSelectedMenu("home")}
-      />
-      <DashboardMenuItem
-        name="Relatórios Financeiros"
-        icon={CreditCard}
-        route="/em-construcao"
-        selected={checkSelected("em-construcao")}
-        onSelect={() => setSelectedMenu("em-construcao")}
-      />
+      {menuItems.map((item) => (
+        <DashboardMenuItem
+          key={item.route}
+          name={item.name}
+          route={item.route}
+          icon={item.icon}
+          selected={pathname.startsWith(item.route)}
+        />
+      ))}
     </section>
   );
 }
@@ -37,30 +33,16 @@ type DashboardMenuItemProps = {
   route: string;
   selected: boolean;
   icon: LucideIcon;
-  onSelect: () => void;
 };
 
-function DashboardMenuItem({
-  name,
-  route,
-  selected,
-  onSelect,
-  icon: Icon,
-}: DashboardMenuItemProps) {
+function DashboardMenuItem({ name, route, selected, icon: Icon }: DashboardMenuItemProps) {
   const itemClass = selected
     ? "bg-brand-tertiary text-white hover:opacity-90"
     : "text-text-secondary hover:bg-gray-100";
 
   return (
     <Link
-      className={[
-        "flex items-center gap-2 rounded-md px-4 py-3",
-        itemClass,
-      ].join(" ")}
-      onClick={onSelect}
-      role="button"
-      tabIndex={0}
-      aria-label={name}
+      className={["flex items-center gap-2 rounded-md px-4 py-3", itemClass].join(" ")}
       href={route}
     >
       <Icon className="size-[18px] shrink-0" aria-hidden />
